added bbr_joiner.py
description: this python script takes as input a csv file of a list of players and their salary information scraped from www.Spotrac.com using the salary_scraper.py file. It then matches each row of this inputted csv file with that players corresponding batting statistics from www.baseballreference.com. Along with saving a csv file to the current working directory (with the name passsed as an argument in the function), the script also prints a list of the players for which it failed to retrieve batting stats.